% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-distances.R
\name{distribution_cdist}
\alias{distribution_cdist}
\alias{distribution_pdist}
\title{Pairwise distances}
\usage{
distribution_cdist(X, Y, method = c("wasserstein", "kolmogorov", "kl", "Lp",
  "moment", "hausdorff", "ky_fan"), ...)

distribution_pdist(X, method = c("wasserstein", "kolmogorov", "kl", "Lp",
  "moment", "hausdorff", "ky_fan"), ...)
}
\arguments{
\item{X, Y}{input data}

\item{method}{used method}

\item{...}{more arguments passed to method}
}
\description{
Compute pairwise distances
}
\examples{
X <- list(rexp(80, rate = 0.2), rexp(80, rate = 0.2))
Y <- list(rexp(120, rate = 0.4), rexp(120, rate = 0.4))
# cdist
distribution_cdist(X, Y, method = "wasserstein")
distribution_cdist(X, Y, method = "kolmogorov")
distribution_cdist(X, Y, method = "kl", k = 5)
distribution_cdist(X, Y, method = "Lp", p = 2)
distribution_cdist(X, Y, method = "moment", p = 1)
distribution_cdist(X, Y, method = "hausdorff", metric = "euclidean")
distribution_cdist(X, Y, method = "ky_fan", metric = "euclidean")

# pdist
distribution_pdist(X, method = "wasserstein")
distribution_pdist(X, method = "kolmogorov")
distribution_pdist(X, method = "kl", k = 5)
distribution_pdist(X, method = "Lp", p = 2)
distribution_pdist(X, method = "moment", p = 1)
distribution_pdist(X, method = "hausdorff", metric = "euclidean")
distribution_pdist(X, method = "ky_fan", metric = "euclidean")
}
\seealso{
\code{\link{wasserstein_dist}}, \code{\link{kl_dist}}, 
\code{\link{Lp_dist}}, \code{\link{moment_dist}}, 
\code{\link{hausdorff_dist}}, \code{\link{ky_fan_dist}}
}
