% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff.R
\name{hausdorff_dist}
\alias{hausdorff_dist}
\title{Hausdorff Distance}
\usage{
hausdorff_dist(X, Y, metric = "euclidean")
}
\arguments{
\item{X, Y}{input data}

\item{metric}{The metric to use}
}
\description{
Compute Hausdorff distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
hausdorff_dist(X, Y, metric = "euclidean")
}
