% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kullback-leibler.R
\name{kl_dist}
\alias{kl_dist}
\title{Kullback-Leibler Distance}
\usage{
kl_dist(X, Y, k = 10)
}
\arguments{
\item{X, Y}{input data}

\item{k}{The number of nearest neighbors to search}
}
\description{
Compute Kullback-Leibler symmetric distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
kl_dist(X, Y, k = 5)
FNN::KL.dist(X, Y, k = 5)[5] + log(120/79) + log(80/119)
}
\references{
S. Boltz, E. Debreuve and M. Barlaud, "kNN-based high-dimensional Kullback-Leibler distance for tracking," Image Analysis for Multimedia Interactive Services, 2007. WIAMIS '07.
}
