% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov.R
\name{kolmogorov_dist}
\alias{kolmogorov_dist}
\title{Kolmogorov Distance}
\usage{
kolmogorov_dist(X, Y)
}
\arguments{
\item{X, Y}{input data}
}
\description{
Compute Kolmogorov distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
kolmogorov_dist(X, Y)
}
