% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky-fan.R
\name{ky_fan_dist}
\alias{ky_fan_dist}
\title{Ky Fan Distance}
\usage{
ky_fan_dist(X, Y, metric = "euclidean")
}
\arguments{
\item{X, Y}{input data}

\item{metric}{The metric to use}
}
\description{
Compute Ky Fan distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
ky_fan_dist(X, Y, metric = "euclidean")
}
