% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lp.R
\name{Lp_dist}
\alias{Lp_dist}
\title{Lp Distance}
\usage{
Lp_dist(X, Y, p = 2)
}
\arguments{
\item{X, Y}{input data}

\item{p}{the power p}
}
\description{
Compute Lp distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
Lp_dist(X, Y, p = 2)
}
