% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment_dist}
\alias{moment_dist}
\title{Moment distance}
\usage{
moment_dist(X, Y, p = 1)
}
\arguments{
\item{X, Y}{input data}

\item{p}{the power p}
}
\description{
Compute moment distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
moment_dist(X, Y, p = 1)
}
