% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein.R
\name{wasserstein_dist}
\alias{wasserstein_dist}
\title{Wasserstein Distance}
\usage{
wasserstein_dist(X, Y, p = 1)
}
\arguments{
\item{X, Y}{input data}

\item{p}{The power p of the pth Wasserstein distance}
}
\description{
Compute Wasserstein distance
}
\examples{
X <- rexp(80, rate = 0.2)
Y <- rexp(120, rate = 0.4)
wasserstein_dist(X, Y)

X <- matrix(X, ncol = 4)
Y <- matrix(Y, ncol = 4)
wasserstein_dist(X, Y)
}
