% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kullback-leibler.R
\name{kl_dist}
\alias{kl_dist}
\title{Kullback-Leibler Distance}
\usage{
kl_dist(X, Y, k = 10)
}
\arguments{
\item{X, Y}{input data}

\item{k}{The number of nearest neighbors to search.}
}
\description{
Compute Kullback-Leibler symmetric distance
}
\references{
S. Boltz, E. Debreuve and M. Barlaud, "kNN-based high-dimensional Kullback-Leibler distance for tracking," Image Analysis for Multimedia Interactive Services, 2007. WIAMIS '07.
}
